requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
bcrypt==3.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23